Start sharing code between GCP and AWS implementations
